use helper for init client + check if we should init sync clients
